Update integration form with documentation link, refactor bulk lines of code into smaller codes
