+++ deepfence_frontend/apps/dashboard/src/features/integrations/pages/AIIntegrationAdd.tsx
@@ -20,7 +20,9 @@ import {
   ModelAddGenerativeAiBedrockIntegrationModelIdEnum,
   ModelAddGenerativeAiOpenAIIntegrationModelIdEnum,
 } from '@/api/generated';
+import { DFLink } from '@/components/DFLink';
 import { SlidingModalHeaderWrapper } from '@/features/common/SlidingModalHeaderWrapper';
+import { IntegrationDocsLinkMap } from '@/features/integrations/components/integration-form/utils';
 import { AI_INTEGRATION_TYPES } from '@/features/integrations/pages/AIIntegrationList';
 import { SuccessModalContent } from '@/features/settings/components/SuccessModalContent';
 import { invalidateAllQueries } from '@/queries';
@@ -188,6 +190,17 @@ const AIIntegrationAdd = () => {
       <SlidingModalContent>
         {!fetcher.data?.success ? (
           <fetcher.Form method="POST" className="flex flex-col gap-8 m-4">
+            <div className="text-p4a text-text-input-value">
+              Get issues remediation. Find out more information by{' '}
+              <DFLink
+                href={IntegrationDocsLinkMap[params.integrationType!]}
+                target="_blank"
+                rel="noreferrer"
+              >
+                reading our documentation
+              </DFLink>
+              .
+            </div>
             {params.integrationType?.length ? (
               <input type="hidden" name="integration_type" value={provider} />
             ) : (
