+++ deepfence_frontend/apps/dashboard/src/features/integrations/pages/IntegrationAdd.tsx
@@ -22,14 +22,17 @@ import {
 } from '@/api/generated';
 import { ErrorStandardLineIcon } from '@/components/icons/common/ErrorStandardLine';
 import { PlusIcon } from '@/components/icons/common/Plus';
-import { integrationTypeToNameMapping } from '@/features/integrations/pages/Integrations';
 import { SuccessModalContent } from '@/features/settings/components/SuccessModalContent';
 import { invalidateAllQueries, queries } from '@/queries';
 import { get403Message, getResponseErrors } from '@/utils/403';
 import { apiWrapper } from '@/utils/api';
 import { getArrayTypeValuesFromFormData } from '@/utils/formData';
 
-import { IntegrationForm, IntegrationType } from '../components/IntegrationForm';
+import {
+  getIntegratinPrettyName,
+  IntegrationType,
+} from '../components/integration-form/utils';
+import { IntegrationForm } from '../components/IntegrationForm';
 import { IntegrationTable } from '../components/IntegrationTable';
 
 export const CLOUD_TRAIL_ALERT = 'CloudTrail Alert';
@@ -639,9 +642,7 @@ const IntegrationAdd = () => {
         >
           <SlidingModalCloseButton />
           <Header
-            title={`Edit Integration: ${
-              integrationTypeToNameMapping[params.integrationType]
-            }`}
+            title={`Edit Integration: ${getIntegratinPrettyName(params.integrationType)}`}
           />
           <IntegrationForm
             integrationType={integrationType}
@@ -660,9 +661,7 @@ const IntegrationAdd = () => {
         >
           <SlidingModalCloseButton />
           <Header
-            title={`Add Integration: ${
-              integrationTypeToNameMapping[params.integrationType]
-            }`}
+            title={`Add Integration: ${getIntegratinPrettyName(params.integrationType)}`}
           />
           <IntegrationForm
             integrationType={integrationType}
